Properly parsing transferLock param in resource_domain.go
The "domain.info" method of the INWX JsonRPC API returns a boolean for the property transferLock. The provider expects a float for some reason. Maybe that was a bug in the API that INWX already has fixed. The type for this property can be confirmed at https://www.inwx.com/en/help/apidoc/f/ch02s09.html#domain.info. 
I have tested it myself and it is indeed a boolean.