Fix matching of records and read of the "name" property
The identifier of inwx_nameserver_record resources is:
DOMAIN_NAME:RO_ID, but what was implemented was RECORD_NAME:RO_ID. That made impossible to use the terraform import command and blocks (v1.5+). 
Also, the value of the record's "name" property was stored in the state as the property "domain". That was fixed as well.